% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_sex_preg}
\alias{age_sex_preg}
\title{Patient initialization table}
\format{
A data frame with 2010 rows and 5 variables:
\describe{
\item{operatingunit}{Country name}
\item{sex}{Sex, Male or Female}
\item{agegroup}{Age Group, "<15", "15+", "<01", "65+", and five-year increments from 1-65}
\item{preg_satus}{Prenancy Status, "pregnant" or "not pregnant"}
\item{prob}{Probability}
}
}
\usage{
age_sex_preg
}
\description{
There are a series of rows that describe a unique and mutually exclusive
categoryof individuals based on age group, sex, and pregnancy status. This
data is transformed from the TX_CURR indicator from the 2023 PEPFAR MSDs.
}
\keyword{datasets}
