% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init-population.R
\name{assign_age}
\alias{assign_age}
\title{Given age_group, chooses age between the boundaries of the age group}
\usage{
assign_age(age_group, anchor_date)
}
\arguments{
\item{age_group}{a vector of age groups}

\item{anchor_date}{YYYY-MM-DD}
}
\value{
list decimal_age, rounded_age, birthday
}
\description{
Given age_group, chooses age between the boundaries of the age group
}
