% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init-interactions.R
\name{assign_interaction_source}
\alias{assign_interaction_source}
\title{Assigns interaction sources to an input interaction df}
\usage{
assign_interaction_source(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
df
}
\description{
Assigns interaction sources to an input interaction df
}
