% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init-interactions.R
\name{assign_interaction_type_source}
\alias{assign_interaction_type_source}
\title{Assigns interaction types and sources given n}
\usage{
assign_interaction_type_source(n)
}
\arguments{
\item{n}{int}
}
\value{
df
}
\description{
Assigns interaction types and sources given n
}
