% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init-interactions.R
\name{assign_pregstatus}
\alias{assign_pregstatus}
\title{Assigns pregnancy status for each interaction}
\usage{
assign_pregstatus(df)
}
\arguments{
\item{df}{an interaction df}
}
\value{
a df
}
\description{
Assigns pregnancy status for each interaction
}
