% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init-interactions.R
\name{calc_age}
\alias{calc_age}
\title{Calculates age give date anchor}
\usage{
calc_age(interaction_dates, dobs)
}
\arguments{
\item{interaction_dates}{a vector of Dates}

\item{dobs}{a vector of dob}
}
\value{
a vector of decimal ages
}
\description{
Calculates age give date anchor
}
