% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init-population.R
\name{calc_birthday}
\alias{calc_birthday}
\title{Calculates birthday}
\usage{
calc_birthday(decimal_age, anchor_date)
}
\arguments{
\item{decimal_age}{dbl}
}
\value{
a Date object birthday
}
\description{
Given decimal age and date of refence (anchor_date), calculates birthday
}
\examples{
calc_birthday(13.4, "20234-01-01")
}
