% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init-interactions.R
\name{calc_pregdate_bounds}
\alias{calc_pregdate_bounds}
\title{Calculates preg date bounds}
\usage{
calc_pregdate_bounds(
  anchor_pregstatus,
  anchor_interaction_date,
  start_gest = 1
)
}
\arguments{
\item{anchor_pregstatus}{a date}

\item{anchor_interaction_date}{a date}

\item{start_gest}{an int}
}
\value{
a list of two vars
}
\description{
Calculates preg date bounds
}
