% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{create_unique_ids}
\alias{create_unique_ids}
\title{Creates unique ids for synthetic dataset}
\usage{
create_unique_ids(n_char, n_ids)
}
\arguments{
\item{n_char}{the number of characters in the ID}

\item{n_ids}{the number of IDs needed}
}
\value{
a vector of IDs
}
\description{
Creates unique ids for synthetic dataset
}
