% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{first_names}
\alias{first_names}
\title{Common first names given country}
\format{
A data frame with 1380 rows and 3 variables:
\describe{
\item{Country}{Country}
\item{First_Name}{First name}
\item{Sex}{Sex, Male or Female}
}
}
\usage{
first_names
}
\description{
This dataset contains common first names generated by ChatGPT
for selected countries.
}
\keyword{datasets}
