% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init-interactions.R
\name{gen_n_interactions}
\alias{gen_n_interactions}
\title{Generates random number of interactions. Not vectorized.}
\usage{
gen_n_interactions(mean = 8, sd = 3, min = 1, max = 15)
}
\arguments{
\item{mean}{int mean number of interactions for normal distribution}

\item{sd}{dbl sd of interactions}

\item{min}{int min interactions}

\item{max}{int max interactions}
}
\value{
n number of interactions that follow the desired normal distribution
}
\description{
Generates random number of interactions. Not vectorized.
}
