% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init-population.R
\name{gen_phonenumber}
\alias{gen_phonenumber}
\title{Generates a vector of random phone numbers (in the Eswatini format)}
\usage{
gen_phonenumber(n_pns)
}
\arguments{
\item{n_pns}{integer; number of phone numbers}
}
\description{
TODO Allow this function to have an input that specifies which country
format to use.
}
