% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_assign-pregdates.R
\name{gen_pregmonth}
\alias{gen_pregmonth}
\title{Generates a random number to represent gestational age in months}
\usage{
gen_pregmonth()
}
\value{
a random integer between 1 and 9
}
\description{
Generates a random number to represent gestational age in months
}
