% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_synthdata.R
\name{gen_synthdata}
\alias{gen_synthdata}
\title{Generate the synthetic dataset}
\usage{
gen_synthdata(country)
}
\arguments{
\item{country}{string}
}
\value{
a synthetic dataset
}
\description{
Generate the synthetic dataset
}
