% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-lab.R
\name{init_lab}
\alias{init_lab}
\title{Initialize the laboratory dataset}
\usage{
init_lab(interactions, country)
}
\arguments{
\item{interactions}{the full interaction data frame}

\item{country}{string}
}
\value{
a dataset of laboratory interaction details
}
\description{
Initialize the laboratory dataset
}
