% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-population.R
\name{init_population}
\alias{init_population}
\title{Initializes the patient data}
\usage{
init_population(country)
}
\arguments{
\item{country}{String}
}
\value{
a df

TODO Change the param, country, to using ISO3 code for reference
TODO Negative probs exist in age_sex_preg; contact Florence to change
}
\description{
Initializes the patient data
}
\examples{
init_population("Eswatini")
}
