% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interaction_type_probs}
\alias{interaction_type_probs}
\title{Interaction type probabilities}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{interaction_type}{Interaction type}
\item{prob}{Probabilities}
}
}
\usage{
interaction_type_probs
}
\description{
Probabilities associated with each interaction type
}
\keyword{datasets}
