% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lab_names}
\alias{lab_names}
\title{Laboratory names}
\format{
A data frame with 1080 rows and 2 variables:
\describe{
\item{country}{Country Name}
\item{lab_name}{Laboratory Name}
}
}
\usage{
lab_names
}
\description{
A data frame of laboratory names
}
\keyword{datasets}
