% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_names}
\alias{last_names}
\title{Common last names given country}
\format{
A data frame with 449 rows and 2 variables:
\describe{
\item{Country}{Country}
\item{Last_Name}{Last name}
}
}
\usage{
last_names
}
\description{
This dataset contains common last names generated by ChatGPT
for selected countries.
}
\keyword{datasets}
