% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_init-interactions.R
\name{update_pregstatus}
\alias{update_pregstatus}
\title{Updates pregnancy status}
\usage{
update_pregstatus(country, df)
}
\arguments{
\item{country}{a string}

\item{df}{a df}
}
\value{
a updated df
}
\description{
Updates pregnancy status
}
